simplified and moved ViewApplyState to runtime/runtime
