+++ core/primitives/src/lib.rs
@@ -11,7 +11,6 @@ pub mod block;
 pub mod block_body;
 pub mod block_header;
 pub mod challenge;
-pub mod compiled_contract;
 pub mod congestion_info;
 pub mod epoch_manager;
 pub mod epoch_sync;
