+++ integration-tests/src/user/runtime_user.rs
@@ -18,11 +18,11 @@ use near_primitives::version::{ProtocolFeature, PROTOCOL_VERSION};
 use near_primitives::views::{
     AccessKeyView, AccountView, BlockView, CallResult, ChunkView, ContractCodeView,
     ExecutionOutcomeView, ExecutionOutcomeWithIdView, ExecutionStatusView,
-    FinalExecutionOutcomeView, FinalExecutionStatus, ViewApplyState, ViewStateResult,
+    FinalExecutionOutcomeView, FinalExecutionStatus, ViewStateResult,
 };
 use near_store::{ShardTries, TrieUpdate};
 use node_runtime::state_viewer::TrieViewer;
-use node_runtime::{ApplyState, Convert, Runtime};
+use node_runtime::{state_viewer::ViewApplyState, ApplyState, Runtime};
 
 use crate::user::{User, POISONED_LOCK_ERR};
 use near_primitives::shard_layout::ShardUId;
@@ -307,7 +307,7 @@ impl User for RuntimeUser {
             epoch_height: apply_state.epoch_height,
             block_timestamp: apply_state.block_timestamp,
             current_protocol_version: PROTOCOL_VERSION,
-            cache: apply_state.cache.map(Convert::convert_back),
+            cache: apply_state.cache,
         };
         result.result = self
             .trie_viewer
